% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rc_project.R
\name{create_rc_project}
\alias{create_rc_project}
\title{This package will create a function called create_rc_project()}
\usage{
create_rc_project(path, ...)
}
\description{
It's callback is at: inst/rstudio/templates/project/create_rc_project.dcf
}
